requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
httpx==0.28.1
python-dotenv==1.0.1
google-search-results==2.4.2
openai==1.58.1
python-multipart==0.0.20
